% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetStockIndex.R
\name{GetStockIndex}
\alias{GetStockIndex}
\title{Index of all Tickers}
\usage{
GetStockIndex(bfull = FALSE)
}
\arguments{
\item{bfull}{boolean defaults to false. If set true, does not filter out based on data issues}
}
\description{
Get an Index of All Tickers listed on the three primary exchanges: NYSE, NASDAQ, and AMEX
}
