% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetStockPrices.R
\name{stock_prices}
\alias{stock_prices}
\alias{GetIntraDay}
\alias{GetTimeSeriesDaily}
\alias{GetTimeSeries}
\title{Stock Price Functions}
\usage{
GetIntraDay(tickers = "MSFT", interval = c("1min", "5min", "15min",
  "30min", "60min"), apikey = "demo")

GetTimeSeriesDaily(tickers = "MSFT", adj = FALSE, apikey = "demo")

GetTimeSeries(tickers = "MSFT", adj = FALSE, period = c("day", "wk",
  "month"), apikey = "demo")
}
\arguments{
\item{tickers}{a vector of one or more tickers to pull data for}

\item{interval}{applicable to IntraDay Prices only. One of: "1min", "5min", "15min", "30min", "60min"}

\item{apikey}{API key from vendor in order to authorize api calls}

\item{adj}{a boolean applicable for weekly, monthly, and daily time horizons. Set TRUE for prices adjusted for splits etc.}

\item{period}{applicable when needing to specify one of "wk" or "month"}
}
\description{
Intraday, daily, then wk/month functions to pull historical or intraday price data
}
\section{Functions}{
\itemize{
\item \code{GetIntraDay}: intraday function. Unique parameter of "interval". See above.

\item \code{GetTimeSeriesDaily}: intraday function. Similar to intraday, has option for compact or full. 
Additionally, set \code{adj = TRUE} when prices should be adjusted (see description)

\item \code{GetTimeSeries}: function for weekly or monthly data specified by 
parameter \code{period} which can be one of: "wk" or "month"
}}

